Move a bunch of helpers to util.py
